+++ lib/api.js
@@ -36,6 +36,7 @@ module.exports = settings => {
   app.use('/me', require('./routers/user'));
   app.use('/invitation', require('./routers/invitation'));
   app.use('/establishment(s)?', require('./routers/establishment'));
+  app.use('/task(s)?', require('./routers/task'));
 
   app.use((req, res, next) => {
     if (res.response) {
