+++ lib/routers/user.js
@@ -18,6 +18,6 @@ router.use((req, res, next) => {
 });
 
 router.use(require('./profile'));
-router.use('/tasks', require('./tasks'));
+router.use('/tasks', require('./task'));
 
 module.exports = router;
