+++ lib/workflow/client.js
@@ -12,10 +12,9 @@ module.exports = settings => {
 
     const client = ApiClient(settings.workflow, { headers });
 
-    req.workflow = params => {
+    req.workflow = (params, path = '/') => {
       if (params.action === 'read') {
-        console.log('attempting to read from workflow cases');
-        return client('/', { method: 'GET', query: params.query });
+        return client(path, { method: 'GET', query: params.query });
       }
 
       return Promise.resolve()
@@ -23,7 +22,7 @@ module.exports = settings => {
           return validate(params);
         })
         .then(() => {
-          return client('/', {
+          return client(path, {
             method: 'POST',
             body: JSON.stringify({ ...params, changedBy: get(req.user, 'profile.id') })
           });
